refactor: separate Effect gateway program from fetch adapter
